Project/LangGraph-Chatbot-main/requirements.txt
streamlit
langgraph
langgraph-checkpoint-postgres
langchain-google-genai
langchain-core
psycopg-binary
psycopg-pool
python-dotenv
